requirements.txt
feedparser
requests
beautifulsoup4
lxml
pandas
PyGithub
openai
playwright